# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080/","description":"Generated server url"}],"paths":{"/tickets/{id}":{"get":{"tags":["tickets-controller"],"operationId":"getTicketById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Ticket"}},"application/xml":{"schema":{"$ref":"#/components/schemas/Ticket"}}}}}},"put":{"tags":["tickets-controller"],"operationId":"checkIn","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/flights/{id}":{"get":{"tags":["flight-controller"],"operationId":"getFlightById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Flight"}},"application/xml":{"schema":{"$ref":"#/components/schemas/Flight"}}}}}},"put":{"tags":["flight-controller"],"operationId":"updateFlight","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateFlightDto"}},"application/xml":{"schema":{"$ref":"#/components/schemas/UpdateFlightDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["flight-controller"],"operationId":"deleteFlight","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/customers/{id}":{"get":{"tags":["customer-controller"],"operationId":"getCustomerById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}},"application/xml":{"schema":{"$ref":"#/components/schemas/Customer"}}}}}},"put":{"tags":["customer-controller"],"operationId":"updateExisting","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCustomerRecordDto"}},"application/xml":{"schema":{"$ref":"#/components/schemas/CreateCustomerRecordDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["customer-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/airports/{id}":{"get":{"tags":["airport-controller"],"operationId":"getAirportById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Airport"}},"application/xml":{"schema":{"$ref":"#/components/schemas/Airport"}}}}}},"put":{"tags":["airport-controller"],"operationId":"updateAirport","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAirportDto"}},"application/xml":{"schema":{"$ref":"#/components/schemas/UpdateAirportDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["airport-controller"],"operationId":"deleteAirport","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/airplanes/{id}":{"get":{"tags":["airplane-controller"],"operationId":"getAirplaneById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Airplane"}},"application/xml":{"schema":{"$ref":"#/components/schemas/Airplane"}}}}}},"put":{"tags":["airplane-controller"],"operationId":"updateAirplane","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AirplaneDto"}},"application/xml":{"schema":{"$ref":"#/components/schemas/AirplaneDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["airplane-controller"],"operationId":"deleteAirplane","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/accounts/confirm/{confirmationTokenId}":{"put":{"tags":["accounts-controller"],"operationId":"confirmAccountRegistration","parameters":[{"name":"confirmationTokenId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/tickets":{"get":{"tags":["tickets-controller"],"operationId":"getAllTickets","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Ticket"}}},"application/xml":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Ticket"}}}}}}},"post":{"tags":["tickets-controller"],"operationId":"createTicket","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseTicketDto"}},"application/xml":{"schema":{"$ref":"#/components/schemas/PurchaseTicketDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Ticket"}}}}}}},"/flights":{"get":{"tags":["flight-controller"],"operationId":"getAllFlights","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Flight"}}}}}}},"post":{"tags":["flight-controller"],"operationId":"createFlight","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFlightDto"}},"application/xml":{"schema":{"$ref":"#/components/schemas/CreateFlightDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int64"}}}}}}},"/customers":{"get":{"tags":["customer-controller"],"operationId":"getAll","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Customer"}}},"application/xml":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Customer"}}}}}}},"post":{"tags":["customer-controller"],"operationId":"createNewCustomer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCustomerAccountDto"}},"application/xml":{"schema":{"$ref":"#/components/schemas/CreateCustomerAccountDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Customer"}}}}}}},"/authenticate":{"post":{"tags":["auth-controller"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/airports":{"get":{"tags":["airport-controller"],"operationId":"getAllAirports","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Airport"}}},"application/xml":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Airport"}}}}}}},"post":{"tags":["airport-controller"],"operationId":"createAirport","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAirportDto"}},"application/xml":{"schema":{"$ref":"#/components/schemas/CreateAirportDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int64"}}}}}}},"/airplanes":{"get":{"tags":["airplane-controller"],"operationId":"getAllAirplanes","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Airplane"}}},"application/xml":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Airplane"}}}}}}},"post":{"tags":["airplane-controller"],"operationId":"createAirplane","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AirplaneDto"}},"application/xml":{"schema":{"$ref":"#/components/schemas/AirplaneDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int64"}}}}}}},"/accounts/test":{"get":{"tags":["accounts-controller"],"operationId":"testMethod","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"UpdateFlightDto":{"type":"object","properties":{"originId":{"type":"string"},"destinationId":{"type":"string"},"airplaneId":{"type":"integer","format":"int64"}}},"CreateCustomerRecordDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"},"addrLine1":{"type":"string"},"addrLine2":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"zipcode":{"type":"string"}}},"UpdateAirportDto":{"type":"object","properties":{"name":{"type":"string"},"streetAddress":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"zipcode":{"pattern":"^\\d{5}(?:[-\\s]\\d{4})?$","type":"string"}}},"Airplane":{"required":["name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"seatConfigurations":{"type":"array","items":{"$ref":"#/components/schemas/SeatConfiguration"}}}},"AirplaneDto":{"required":["name","seatConfigurations"],"type":"object","properties":{"name":{"type":"string"},"seatConfigurations":{"type":"array","items":{"$ref":"#/components/schemas/SeatConfiguration"}}}},"SeatConfiguration":{"required":["numRows","numSeatsPerRow"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"airplane":{"$ref":"#/components/schemas/Airplane"},"seatClass":{"type":"string","enum":["ECONOMY","BUSINESS"]},"numRows":{"type":"integer","format":"int32"},"numSeatsPerRow":{"type":"integer","format":"int32"}}},"PurchaseTicketDto":{"required":["flightId","purchaserId","tickets"],"type":"object","properties":{"purchaserId":{"type":"integer","format":"int64"},"flightId":{"type":"integer","format":"int64"},"tickets":{"type":"array","items":{"$ref":"#/components/schemas/TicketItem"}}}},"TicketItem":{"required":["passengerName","seatClass","seatNumber"],"type":"object","properties":{"seatClass":{"type":"string"},"seatNumber":{"type":"string"},"passengerName":{"type":"string"}}},"Ticket":{"required":["flightId","passengerName","purchaserId","seatClass","seatNumber","status"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"flightId":{"type":"integer","format":"int64"},"purchaserId":{"type":"integer","format":"int64"},"passengerName":{"type":"string"},"seatClass":{"type":"string"},"seatNumber":{"type":"string"},"status":{"type":"string","enum":["PURCHASED","CHECKED_IN","CANCELLED","REFUNDED"]},"customerId":{"type":"integer","format":"int64"},"passenger":{"type":"string"}}},"CreateFlightDto":{"required":["airplaneId","baseSeatPrice","destinationId","originId"],"type":"object","properties":{"originId":{"type":"string"},"destinationId":{"type":"string"},"airplaneId":{"type":"integer","format":"int64"},"baseSeatPrice":{"minimum":0,"type":"number"}}},"CreateCustomerAccountDto":{"required":["addrLine1","city","email","firstName","lastName","password","state","zipcode"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":128,"minLength":10,"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[!@#$%^&*-_=+,.?])[A-Za-z\\d!@#$%^&*-_=+,.?]{10,128}$","type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"phoneNumber":{"pattern":"^\\d{3}-\\d{3}-\\d{4}$","type":"string"},"addrLine1":{"type":"string"},"addrLine2":{"type":"string"},"city":{"type":"string"},"state":{"maxLength":2,"minLength":2,"type":"string"},"zipcode":{"pattern":"^\\d{5}(?:[-\\s]\\d{4})?$","type":"string"},"accountDto":{"$ref":"#/components/schemas/CreateUserAccountDto"},"recordDto":{"$ref":"#/components/schemas/CreateCustomerRecordDto"}}},"CreateUserAccountDto":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":128,"minLength":10,"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[!@#$%^&*-_=+,.?])[A-Za-z\\d!@#$%^&*-_=+,.?]{10,128}$","type":"string"}}},"Address":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"cardinality":{"type":"integer","format":"int32"},"line1":{"type":"string"},"line2":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"zipcode":{"type":"string"}}},"Customer":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"addresses":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Address"}},"phoneNumber":{"type":"string"},"loyaltyPoints":{"type":"integer","format":"int32"},"paymentMethods":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/PaymentMethod"}}}},"PaymentMethod":{"required":["accountNum","ownerId"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"ownerId":{"type":"integer","format":"int64"},"accountNum":{"type":"string"},"notes":{"type":"string"}}},"AuthDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"CreateAirportDto":{"required":["city","iataId","name","state","streetAddress","zipcode"],"type":"object","properties":{"iataId":{"type":"string"},"name":{"type":"string"},"streetAddress":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"zipcode":{"pattern":"^\\d{5}(?:[-\\s]\\d{4})?$","type":"string"}}},"Airport":{"required":["city","iataId","name","state","streetAddress","zipcode"],"type":"object","properties":{"iataId":{"type":"string"},"name":{"type":"string"},"streetAddress":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"zipcode":{"type":"string"}}},"Flight":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"origin":{"$ref":"#/components/schemas/Airport"},"destination":{"$ref":"#/components/schemas/Airport"},"airplane":{"$ref":"#/components/schemas/Airplane"},"seats":{"type":"array","items":{"$ref":"#/components/schemas/Seat"}}}},"Seat":{"required":["price","seatColumn","seatRow"],"type":"object","properties":{"id":{"type":"string"},"flight":{"$ref":"#/components/schemas/Flight"},"seatRow":{"type":"integer","format":"int32"},"seatColumn":{"type":"string"},"seatClass":{"type":"string","enum":["ECONOMY","BUSINESS"]},"seatStatus":{"type":"string","enum":["AVAILABLE","SOLD","HELD","BLOCKED"]},"price":{"type":"number"}}}}}}